requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy==1.26.4
tensorflow==2.19.0  # For .keras Keras/TensorFlow models
keras==3.5.0

# community feature
sqlalchemy
psycopg2-binary    # Postgres driver
boto3              # AWS S3 client
pydantic
python-dotenv      # Environment variables
